adding counts to department page
