+++ frontends/mit-learn/src/common/utils.ts
@@ -1,3 +1,4 @@
+import { ChannelCounts } from "api/v0"
 const getSearchParamMap = (urlParams: URLSearchParams) => {
   const params: Record<string, string[] | string> = {}
   for (const [key] of urlParams.entries()) {
@@ -6,4 +7,27 @@ const getSearchParamMap = (urlParams: URLSearchParams) => {
   return params
 }
 
-export { getSearchParamMap }
+const aggregateProgramCounts = (
+  groupBy: string,
+  data: Array<ChannelCounts>,
+): Record<string, number> => {
+  return Object.fromEntries(
+    Object.entries(data).map(([_key, value]) => {
+      console.log("value", value)
+      return [value[groupBy], value.counts.programs]
+    }),
+  )
+}
+
+const aggregateCourseCounts = (
+  groupBy: string,
+  data: Array<ChannelCounts>,
+): Record<string, number> => {
+  return Object.fromEntries(
+    Object.entries(data).map(([_key, value]) => {
+      return [value[groupBy], value.counts.courses]
+    }),
+  )
+}
+
+export { getSearchParamMap, aggregateProgramCounts, aggregateCourseCounts }
